Arbitrage Bot/requirements.txt
ccxt==4.4.96
websockets==15.0.1
python-dotenv==1.1.1
pandas==2.3.1
numpy==2.3.2
requests==2.32.4
aiohttp==3.12.14
asyncio==3.4.3
psutil==7.0.0
flask==3.1.1
flask-cors==6.0.1

